+++ NEWS.md
@@ -1,4 +1,4 @@
-# mlr3tuning 0.11.0.9000
+# mlr3tuning 0.12.0
 
 * feat: `as_search_space()` function to create search spaces from `Learner` and
   `ParamSet` objects. Allows to pass `TuningSpace` objects as `search_space` in